backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
docx2txt==0.9
faiss-cpu==1.11.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
numpy==2.2.5
openai==1.78.0
packaging==25.0
pydantic==2.11.4
pydantic_core==2.33.2
pypdf==5.4.0
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
tenacity==8.2.3
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
